Relocate JavaScript tests into the npm directory
