--- __tests__/unit/compatibility/jquery.js
@@ -1,27 +1,0 @@
-/* global given */
-
-require('@cocooned/src/javascripts/cocooned')
-const { asAttribute } = require('@cocooned/tests/support/helpers')
-
-describe('A Cocooned setup', () => {
-  given('template', () => `
-    <section>
-      <div>
-        <a class="cocooned-add" href="#"
-           data-associations="items"
-           data-association-insertion-template="${asAttribute(given.insertionTemplate)}">Add</a>
-      </div>
-    </section>
-  `)
-  given('insertionTemplate', () => '<div class="cocooned-item"></div>')
-  given('container', () => document.querySelector('section'))
-
-  beforeEach(() => {
-    document.body.innerHTML = given.template
-    $(given.container).cocooned()
-  })
-
-  it('is instanciated as a Cocooned instance', () => {
-    expect(given.container).toHaveClass('cocooned-container')
-  })
-})
