+++ npm/__tests__/support/jquery.js
